# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GymAssist API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Workout {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  goal: string;
  durationMinutes: number;
  difficulty: string;
  completed: boolean;
  /** @nullable */
  scheduledAt?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateWorkoutBody {
  name: string;
  description?: string;
  goal: string;
  durationMinutes: number;
  difficulty: string;
  scheduledAt?: string;
}

export interface UpdateWorkoutBody {
  name?: string;
  description?: string;
  goal?: string;
  durationMinutes?: number;
  difficulty?: string;
  completed?: boolean;
  scheduledAt?: string;
}

export interface Exercise {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  muscleGroup: string;
  difficulty: string;
  /** @nullable */
  equipment?: string | null;
  /** @nullable */
  instructions?: string | null;
  /** @nullable */
  sets?: number | null;
  /** @nullable */
  reps?: number | null;
  /** @nullable */
  durationSeconds?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateExerciseBody {
  name: string;
  description?: string;
  muscleGroup: string;
  difficulty: string;
  equipment?: string;
  instructions?: string;
  sets?: number;
  reps?: number;
  durationSeconds?: number;
}

export interface UpdateExerciseBody {
  name?: string;
  description?: string;
  muscleGroup?: string;
  difficulty?: string;
  equipment?: string;
  instructions?: string;
  sets?: number;
  reps?: number;
  durationSeconds?: number;
}

export interface HealthMetric {
  id: number;
  type: string;
  value: number;
  unit: string;
  /** @nullable */
  notes?: string | null;
  loggedAt: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateHealthMetricBody {
  type: string;
  value: number;
  unit: string;
  notes?: string;
  loggedAt?: string;
}

export interface UpdateHealthMetricBody {
  type?: string;
  value?: number;
  unit?: string;
  notes?: string;
  loggedAt?: string;
}

export interface ChatMessageBody {
  message: string;
}

export interface ChatResponse {
  reply: string;
  messageId: number;
}

export interface ChatMessage {
  id: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface DashboardSummary {
  totalWorkouts: number;
  completedWorkouts: number;
  totalExercises: number;
  /** @nullable */
  latestWeight?: number | null;
  /** @nullable */
  latestWeightUnit?: string | null;
  weeklyWorkoutsCount: number;
  /** @nullable */
  averageDurationMinutes?: number | null;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  timestamp: string;
}

export interface MetricDataPoint {
  loggedAt: string;
  value: number;
}

export interface MetricTrend {
  type: string;
  unit: string;
  data: MetricDataPoint[];
}

export type ListExercisesParams = {
  muscleGroup?: string;
  difficulty?: string;
};

export type ListHealthMetricsParams = {
  type?: string;
  limit?: number;
};