# lib/api-zod/src/generated/types/createExerciseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GymAssist API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateExerciseBody {
  name: string;
  description?: string;
  muscleGroup: string;
  difficulty: string;
  equipment?: string;
  instructions?: string;
  sets?: number;
  reps?: number;
  durationSeconds?: number;
}