# lib/api-zod/src/generated/types/createWorkoutBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GymAssist API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateWorkoutBody {
  name: string;
  description?: string;
  goal: string;
  durationMinutes: number;
  difficulty: string;
  scheduledAt?: string;
}