# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GymAssist API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalWorkouts: number;
  completedWorkouts: number;
  totalExercises: number;
  /** @nullable */
  latestWeight?: number | null;
  /** @nullable */
  latestWeightUnit?: string | null;
  weeklyWorkoutsCount: number;
  /** @nullable */
  averageDurationMinutes?: number | null;
}