# lib/api-zod/src/generated/types/healthMetric.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GymAssist API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HealthMetric {
  id: number;
  type: string;
  value: number;
  unit: string;
  /** @nullable */
  notes?: string | null;
  loggedAt: string;
  createdAt: string;
  updatedAt: string;
}