# lib/api-zod/src/generated/types/updateWorkoutBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GymAssist API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateWorkoutBody {
  name?: string;
  description?: string;
  goal?: string;
  durationMinutes?: number;
  difficulty?: string;
  completed?: boolean;
  scheduledAt?: string;
}