# lib/api-zod/src/generated/types/workout.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GymAssist API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Workout {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  goal: string;
  durationMinutes: number;
  difficulty: string;
  completed: boolean;
  /** @nullable */
  scheduledAt?: string | null;
  createdAt: string;
  updatedAt: string;
}